Update patch set 1
Patch Set 1:

(5 comments)

> Guy - would you add the Apache license header? https://wiki.acumos.org/display/AC/License+Templates#LicenseTemplates-Python,Shell,YAML
 > Does this code need a requirements.txt file?
 > possibly a README?

This is example code from the tutorial document. My goal is to keep this as clean and clear as possible. I could surely add the boilerplate license stuff, but it does serve to junk up the code. Note that Iris_sklearn.py is lifted from https://pypi.org/project/acumos/ and no licensing junk appears there.

The tutorial document serves as a README. Can I put a MS Word doc into this repo?

As for a requirements.txt--maybe. Again, this is not code we are distributing for people to run right out of the box. It's just example code. I put it into this repo because I was asked to do so.

Patch-set: 1