Update patch set 1
Patch Set 1:

> 1) all files submitted to gerrit need to be licensed as long as the
 > language supports comments (Acumos-203)
 > 2)  https://pypi.org/project/acumos/ is dynamically generated from
 > the acumos-python-client (see setup.py L32037 in that repo), which
 > contains the relevant license headers
 > 3) in order to be part of the automated documentation
 > (docs.acumos.org), docs should be in RST format; if you choose to
 > add it to gerrit, the Soup to Nuts guide needs to be converted to
 > RST and then added to this repo in a docs folder, which should
 > include an index.rst file; or the Soup to Nuts page can be left on
 > the wiki (regardless, I will move your page to a better location on
 > the wiki)

If I were to add the model.proto file or the generated python code, they would not have any licensing. Both of these are auto-generated and are not user-created.

I was thinking of adding these to this repo to help as examples, but I cannot really go adding comments to auto-generated files (even if they support comments).

So I guess I cannot add these to this repo, yes?

Patch-set: 1