Update patch set 1
Patch Set 1:

As a readme, I will just point to the wiki or the docs thing.

It's kind of a bad thing to have multiple versions of the same document, because they will get out of sync.

I find RST kind of a pain to work with. There is a tool to convert MS Word to RST but it generally does a poor job and needs by-hand editing afterwards to look decent.

Patch-set: 1