Update patch set 11
Patch Set 11:

Guys - please merge if possible, this has been out for some time now. And it's clear that the JWT auth method is not going away soon. I'm investigating the other options to support automated auth but in the meantime this works, and is a valuable set of tools that need to be in the repo. Thanks for your help.

Patch-set: 11
Reviewer: Gerrit User 93 <93@655300a2-7a56-4504-9153-881edd907bea>
Label: Code-Review=0