Update patch set 1
Patch Set 1: Code-Review+1

Since you are changing the import statement, I recommend that someone test this to see that it does not break the example. But that would be optional, depending upon the status of this demo model (if it's a lot of work to test, and not one of our main demo models, I would skip the testing). If you not expecting to test this, I can merge it. Let me know when you are ready for a merge.

Patch-set: 1