Update patch set 1
Patch Set 1:

Sigh.  Hoping to move faster than this and pick things like license on next batch of changes.

All images and videos are needed for demos across different model purposes.  Each file is far less than the 5MB limit, so shouldn't be any qualms there.  This demo app is the evolution of the previous project-specific (and limited) versions. 

Some of the commented out code is left as examples for future development and expansion by others, so mixed opinion for absolute removal.

Patch-set: 1